Added Toggle for Colors
