+++ public/rootStyles.scss
@@ -27,9 +27,13 @@ $background: $lightestCultured;
 }
 h1 {
   font-size: 3vw;
-  background-color: $lightestCultured;
   text-align: center;
   border-bottom: 1px solid currentColor;
+  cursor: pointer;
+  transition: color 150ms;
+  &:hover {
+    color:$lightSpanishPink;
+  }
 }
 .logo {
   display: flex;
